Fix job imports and update zip extraction
